refactor: update cheese detail and add functionality to add cheese with date and notes

+++ src/app/pages/my-cheeses/cheese-detail/cheese-detail.page.ts
@@ -1,23 +1,22 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonModule, Location } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { IonContent, IonHeader, IonTitle, IonToolbar, IonSpinner, IonAccordion, IonItem, IonLabel, IonAccordionGroup } from '@ionic/angular/standalone';
-import { ActivatedRoute, ParamMap, Router } from '@angular/router';
+import { IonContent, IonHeader, IonTitle, IonToolbar, IonSpinner, IonAccordion, IonItem, IonLabel, IonAccordionGroup, IonIcon, IonButton } from '@ionic/angular/standalone';
+import { ActivatedRoute, ParamMap, Router, RouterLink } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { Cheese } from 'src/app/interfaces/cheese';
 import { CheeseService } from 'src/app/services/cheese.service';
 import { MenuComponent } from "src/app/components/menu/menu.component";
 
-
 import { addIcons } from 'ionicons';
-import { arrowBackCircleOutline, caretDownCircle, caretDownCircleOutline } from 'ionicons/icons';
+import { arrowBackCircleOutline, caretDownCircle, caretDownCircleOutline, createOutline } from 'ionicons/icons';
 import { CheeseDetailComponent } from 'src/app/components/cheese-detail/cheese-detail.component';
 @Component({
   selector: 'app-cheese-detail-page',
   templateUrl: './cheese-detail.page.html',
   styleUrls: ['./cheese-detail.page.scss'],
   standalone: true,
-  imports: [IonContent, IonHeader, IonTitle, IonToolbar, CommonModule, FormsModule, MenuComponent, CheeseDetailComponent, IonSpinner, IonAccordion, IonItem, IonLabel, IonAccordionGroup]
+  imports: [IonContent, IonHeader, IonTitle, IonToolbar, CommonModule, FormsModule, MenuComponent, CheeseDetailComponent, IonSpinner, IonAccordion, IonItem, IonLabel, IonAccordionGroup, IonIcon, IonButton]
 })
 export class CheeseDetailPage implements OnInit {
 
@@ -26,10 +25,10 @@ export class CheeseDetailPage implements OnInit {
   isLoading: boolean = true;
 
   private routeSub!: Subscription;
-  constructor(private route: ActivatedRoute, private cheeseService: CheeseService, private location: Location, private router: Router) {
+  constructor(private route: ActivatedRoute, private cheeseService: CheeseService, private router: Router) {
     // Add icons to the IonIcon component
     addIcons({
-      arrowBackCircleOutline: arrowBackCircleOutline, caretDownCircle
+      arrowBackCircleOutline: arrowBackCircleOutline, caretDownCircle, createOutline
     })
    }
 ngOnInit(): void {
@@ -48,7 +47,7 @@ loadCheese(): void {
     next: (response: { msg: string; cheese: Cheese }): void => {
       this.cheese = response.cheese;
       this.isLoading = false;
-      console.log('Formatge carregat:', this.cheese);
+      console.log('Formatge carregat:', this.cheese._id);
     },
     error: (error: any): void => {
       console.error('Error carregant el formatge:', error);
@@ -56,15 +55,19 @@ loadCheese(): void {
     }
   });
   }
- deleteCheese(id: string): void {
-    this.cheeseService.deleteCheese(id).subscribe({
-      next: (response) => {
-        console.log('Formatge eliminat:', response);
-        this.router.navigate(['/my-cheeses']);
-      },
-      error: (error) => {
-        console.error('Error eliminant el formatge:', error);
-      }
-    });
+async deleteCheese(id: string): Promise<void> {
+  const confirmed = window.confirm('are you sure you want to delete this cheese?');
+  if (!confirmed) {
+    return;
   }
+  this.cheeseService.deleteCheese(id).subscribe({
+    next: (response) => {
+      console.log('Formatge eliminat:', response);
+      this.router.navigate(['/my-cheeses']);
+    },
+    error: (error) => {
+      console.error('Error eliminant el formatge:', error);
+    }
+  });
+}
 }
